Reworked UI sizes
Store UI size as float with 1. as default. Use float factor to resize UI.

+++ ProxyStyle.h
@@ -27,7 +27,7 @@ public:
 private:
     Q_DISABLE_COPY(ProxyStyle)
 
-    static int actualUisize_;
+    static float actualUisize_;
 };
 
 #endif // PROXYSTYLE_H
